Add support for custom ranking
