+++ .gitignore
@@ -3,3 +3,4 @@ dist
 package-lock.json
 node_modules
 src/db-tests
+src/custom
